requirements.txt
Spire.Doc==12.4.0
